+++ packages/upload-sdk/tsup.config.ts
@@ -2,15 +2,17 @@ import { defineConfig } from "tsup"
 import { inlineWorkerPlugin } from "./plugins/inline-worker"
 import path from "path"
 
+const isDev = process.env.NODE_ENV === "development"
+
 export default defineConfig(_options => ({
     entry: ["src/index.ts"],
     splitting: false,
-    sourcemap: false,
+    sourcemap: isDev,
     clean: true,
     outDir: "es",
     format: ["esm"],
     dts: true,
-    minify: true,
+    minify: !isDev,
     esbuildPlugins: [inlineWorkerPlugin()],
 
     esbuildOptions(options, _context) {